pyproject.toml
[tool.poetry]
name = "cookbook-demo"
version = "0.1.0"
description = ""
authors = ["Alex W <alex@lightsonsoftware.com>"]
readme = "README.md"
packages = [{include = "cookbook_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
nbconvert = "^7.16.4"
nbformat = "^5.10.4"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
